decomp shared logic in collection tabs
